+++ wdl-lsp/CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-* Fixed an issue where the LSP would error out when a file URI was received instead of a directory URI
+* Fixed an issue where the LSP was expecting a directory URI instead of file URI ([#362](https://github.com/stjude-rust-labs/wdl/pull/362)).
 
 ## 0.6.0 - 01-17-2025